refactor: Refactored our service.ts to a new package structure with defined responsibility in each file
