+++ my-medusa-store/src/api/admin/embeddings/[product_id]/route.ts
@@ -3,7 +3,7 @@ import type {
   MedusaResponse,
 } from "@medusajs/framework/http";
 import { ELASTICSEARCH_MODULE } from "../../../../modules/elasticsearch";
-import ElasticsearchModuleService from "../../../../modules/elasticsearch/service";
+import ElasticsearchModuleService from "../../../../modules/elasticsearch/services/main";
 
 // GET embedding by product ID
 export const GET = async (
