+++ my-medusa-store/src/modules/elasticsearch/README.md
@@ -83,7 +83,7 @@ The module service is available via dependency injection:
 
 ```typescript
 import { ELASTICSEARCH_MODULE } from "../../modules/elasticsearch";
-import ElasticsearchModuleService from "../../modules/elasticsearch/service";
+import ElasticsearchModuleService from "../../modules/elasticsearch/services/main";
 
 // In a workflow step
 const elasticsearchService: ElasticsearchModuleService =
