+++ my-medusa-store/src/modules/elasticsearch/index.ts
@@ -1,5 +1,5 @@
 import { Module } from "@medusajs/framework/utils";
-import ElasticsearchModuleService from "./service";
+import ElasticsearchModuleService from "./services/main";
 
 export const ELASTICSEARCH_MODULE = "elasticsearch";
