+++ my-medusa-store/src/modules/elasticsearch/types/index.ts
