+++ my-medusa-store/src/scripts/worker.ts
@@ -1,6 +1,6 @@
 import 'dotenv/config';
 import { ELASTICSEARCH_MODULE } from "../modules/elasticsearch";
-import ElasticsearchModuleService from "../modules/elasticsearch/service";
+import ElasticsearchModuleService from "../modules/elasticsearch/services/main";
 
 console.log("🚀 Product Embedding Worker Starting...");
