+++ my-medusa-store/src/workflows/search/steps/search-index.ts
@@ -1,6 +1,6 @@
 import { createStep, StepResponse } from "@medusajs/framework/workflows-sdk";
 import { ELASTICSEARCH_MODULE } from "../../../modules/elasticsearch";
-import ElasticsearchModuleService from "../../../modules/elasticsearch/service";
+import ElasticsearchModuleService from "../../../modules/elasticsearch/services/main";
 import { EmbeddingResult } from "../../../lib/embedding-client";
 import { SearchMode } from "../../../modules/elasticsearch/types";
